backend/pyproject.toml
[tool.poetry]
name = "stock-rags"
version = "0.1.0"
description = "GraphRAG Stock Report Analysis Service"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

# Database
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
neo4j = "^5.16.0"
qdrant-client = "^1.7.0"
redis = "^5.0.1"

# LLM & AI
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-anthropic = "^0.3.0"
langchain-google-genai = "^2.0.0"
langgraph = "^0.2.0"
openai = "^1.50.0"
anthropic = "^0.39.0"
google-generativeai = "^0.8.0"

# PDF Processing
pypdf = "^3.17.4"
pdfplumber = "^0.10.3"

# Task Queue
celery = {extras = ["redis"], version = "^5.3.4"}

# Utilities
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
httpx = "^0.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
black = "^23.12.1"
ruff = "^0.1.11"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
